Allow for logging out with HTTP-POST
Beyond setting a default logout binding, there doesn't appear to
be a way to specify on a LogoutRequest which binding a IdP should
reply to. pysaml2's example IdP just resuses the binding that the
request came in on. So if the IdP only allows for HTTP-POST requests
it'll send a HTTP-POST reply. Therefore it makes sense to support
both HTTP-REDIRECT and HTTP-POST methods for logging out.